requirements.txt
snakebite-py3==3.0.5
apache-airflow-providers-celery==3.1.0
zope.component==6.0
zope.event==4.5.0
zope.hookable==5.4
zope.interface==5.3.0
aiohttp==3.7.3
celery==5.2.3
Scrapy==2.9.0
constantly == 15.1
incremental == 21.3.0
Automat == 0.8.0
hyperlink == 17.1.1
attrs == 21.3.0
twilio
scrapy_cloudflare_middleware
cfscrape>=1.9.4
pykafka
apache-airflow-providers-celery
fastapi
pykafka == 2.1.0
faker
pymongo[srv]


 